requirements.txt
Django==4.2
django-extensions
djangorestframework
django-q
requests
coverage